# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PoupaMais - Personal Finance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/onboarding:
    post:
      operationId: completeOnboarding
      tags: [users]
      summary: Complete onboarding setup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "200":
          description: Onboarding completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /transactions:
    get:
      operationId: getTransactions
      tags: [transactions]
      summary: List all transactions
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Transaction updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
      required:
        - success
...
    User:
      type: object
      required: [id, name, email, currency, language, isPremium, onboardingCompleted, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        currency:
          type: string
        language:
          type: string
        isPremium:
          type: boolean
        premiumExpiresAt:
          type: ["string", "null"]
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
    ProfileUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        currency:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
    OnboardingInput:
      type: object
      required: [currency, initialGoalAmount, favoriteCategories]
      properties:
        currency:
          type: string
        initialGoalAmount:
          type: number
        initialGoalName:
          type: ["string", "null"]
        favoriteCategories:
          type: array
          items:
            type: string
    Transaction:
      type: object
      required: [id, userId, type, amount, description, date, categoryId, categoryName, categoryColor, categoryIcon, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        description:
          type: string
        date:
          type: string
        time:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryColor:
          type: string
        categoryIcon:
          type: string
        isRecurring:
          type: boolean
        recurringPeriod:
          type: ["string", "null"]
        installments:
          type: ["integer", "null"]
        installmentNumber:
          type: ["integer", "null"]
        cardId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    TransactionInput:
      type: object
      required: [type, amount, description, date, categoryId]
      properties:
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
          minimum: 0.01
        description:
          type: string
          minLength: 1
        date:
          type: string
        time:
          type: ["string", "null"]
        categoryId:
          type: integer
        isRecurring:
          type: ["boolean", "null"]
        recurringPeriod:
          type: ["string", "null"]
          enum: [weekly, monthly, yearly, null]
        installments:
          type: ["integer", "null"]
        cardId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    TransactionUpdate:
      type: object
      properties:
        type:
          type: ["string", "null"]
          enum: [income, expense, null]
        amount:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        time:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        isRecurring:
          type: ["boolean", "null"]
        recurringPeriod:
          type: ["string", "null"]
        installments:
          type: ["integer", "null"]
        cardId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
...