# lib/api-zod/src/generated/types/transactionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PoupaMais - Personal Finance API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionUpdateType } from './transactionUpdateType';

export interface TransactionUpdate {
  /** @nullable */
  type?: TransactionUpdateType;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  time?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  isRecurring?: boolean | null;
  /** @nullable */
  recurringPeriod?: string | null;
  /** @nullable */
  installments?: number | null;
  /** @nullable */
  cardId?: number | null;
  /** @nullable */
  notes?: string | null;
}